Print Promotion Name If Price Dropped
Prints the promotion name (if there is one) when there is a price drop.
There almost always is a promotion name, but usually it is meaningless. However, if the price is lower, we print it so you might feel a little better about catching  deal and knowing the reason. Many times these deals are specific to you: age, C&A status, cabin selection. Other trackers like royalpricetracker will not find these!